Restructured the RoleDeterminer a bit to reflect v12's features more.

+++ src/Utility/Muted.js
@@ -9,7 +9,7 @@ module.exports = BotModule.extend({
     initialize: function (dependencyGraph) {
         this._super(dependencyGraph);
         this.discordClient.on('message', message => {
-            if (RoleDeterminer.isMuted(message.member)) {
+            if (message.member && RoleDeterminer.isMuted(message.member)) {
                 message.delete();
             }
         });
